// frontend/features/api/model/updateSupplierRequest.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * Petrol Integrity System API Documentation
 * OpenAPI spec version: 0.0.1
 */

export interface UpdateSupplierRequest {
  /** @maxLength 255 */
  name?: string;
  /**
   * @maxLength 255
   * @nullable
   */
  email?: string | null;
  /**
   * @maxLength 20
   * @nullable
   */
  phone?: string | null;
}